changes to include the distributed operations in the aten_ops lib
