+++ py/torch_tensorrt/dynamo/conversion/impl/__init__.py
@@ -13,6 +13,7 @@ from torch_tensorrt.dynamo.conversion.impl import (
     full,
     grid,
     matmul,
+    nccl_ops,
     normalization,
     pad,
     permutation,
